chore: migrate backend index to TypeScript and add basic types
