+++ backend/src/types/types.ts
@@ -1,0 +1,14 @@
+export interface Candle {
+  open: number;
+  high: number;
+  low: number;
+  close: number;
+}
+
+export type OHLCData = Record<string, Record<string, Candle>>;
+
+export interface WeatherEvent {
+  city: string;
+  timestamp: string;
+  temperature: number;
+}
